right before changing image spawning method

+++ ImageWorld/Assets/SaveStage.cs
@@ -7,7 +7,9 @@ using UnityEngine.SceneManagement;
 
 public class SaveStage : MonoBehaviour
 {
-    GameObject Images;
+    public bool isHost;
+
+    public GameObject Images;
     public string worldName;
 
     private void Awake()
@@ -25,6 +27,8 @@ public class SaveStage : MonoBehaviour
         }
     }
 
+    
+
     // Update is called once per frame
     void Update()
     {
@@ -36,36 +40,40 @@ public class SaveStage : MonoBehaviour
 
     public void SaveWorld()
     {
-        if (worldName == "" || worldName==null) { return; }
-        if (Directory.Exists("C:/CollageWorld/SavedInfo/" + $"{worldName}"))
-        if (Directory.Exists("C:/CollageWorld/SavedInfo/" + $"{worldName}"))
-        { Directory.Delete("C:/CollageWorld/SavedInfo/" + $"{worldName}", true); }
-
-        Directory.CreateDirectory("C:/CollageWorld/SavedInfo/" + $"{worldName}");
-
-        if (File.Exists("C:/CollageWorld/SavedInfo/" + $"{worldName}/{worldName}.txt"))
-        { File.Delete("C:/CollageWorld/SavedInfo/" + $"{worldName}/{worldName}.txt"); }
-
-        StreamWriter writer = File.AppendText("C:/CollageWorld/SavedInfo/" + $"{worldName}/{worldName}.txt");
-
-        writer.WriteLine($"{Images.transform.childCount}");
-        writer.WriteLine($"------------------------------------------");
-
-        foreach (TextureHavenScript i in Images.GetComponentsInChildren<TextureHavenScript>())
+        Images = GameObject.Find("ALL_IMAGES");
+        if (SceneManager.GetActiveScene().buildIndex == 2 || isHost)
         {
-            byte[] savedImage = i.editTex.EncodeToPNG();
-            File.WriteAllBytes("C:/CollageWorld/SavedInfo/" + $"{worldName}/" + i.IMAGEID.ToString() + ".png", savedImage);
-            writer.WriteLine($"{i.IMAGEID}");
-            writer.WriteLine($"POS: \n{i.transform.position.x},{i.transform.position.y},{i.transform.position.z}");
-            writer.WriteLine($"ROT: \n{i.transform.rotation.eulerAngles.x},{i.transform.rotation.eulerAngles.y},{i.transform.rotation.eulerAngles.z}");
-            writer.WriteLine($"SIZ: \n{i.transform.localScale.x},{i.transform.localScale.y},{i.transform.localScale.z}");
-            writer.WriteLine($"COL: \n{i.currentColor.r},{i.currentColor.g},{i.currentColor.b},{i.currentColor.a}");
-            writer.WriteLine($"SKT: \n{i.sketchified.ToString()[0]}");
+            if (worldName == "" || worldName == null) { return; }
+            if (Directory.Exists("C:/CollageWorld/SavedInfo/" + $"{worldName}"))
+                if (Directory.Exists("C:/CollageWorld/SavedInfo/" + $"{worldName}"))
+                { Directory.Delete("C:/CollageWorld/SavedInfo/" + $"{worldName}", true); }
+
+            Directory.CreateDirectory("C:/CollageWorld/SavedInfo/" + $"{worldName}");
+
+            if (File.Exists("C:/CollageWorld/SavedInfo/" + $"{worldName}/{worldName}.txt"))
+            { File.Delete("C:/CollageWorld/SavedInfo/" + $"{worldName}/{worldName}.txt"); }
+
+            StreamWriter writer = File.AppendText("C:/CollageWorld/SavedInfo/" + $"{worldName}/{worldName}.txt");
+
+            writer.WriteLine($"{Images.transform.childCount}");
             writer.WriteLine($"------------------------------------------");
+
+            foreach (TextureHavenScript i in Images.GetComponentsInChildren<TextureHavenScript>())
+            {
+                byte[] savedImage = i.editTex.EncodeToPNG();
+                File.WriteAllBytes("C:/CollageWorld/SavedInfo/" + $"{worldName}/" + i.IMAGEID.ToString() + ".png", savedImage);
+                writer.WriteLine($"{i.IMAGEID}");
+                writer.WriteLine($"POS: \n{i.transform.position.x},{i.transform.position.y},{i.transform.position.z}");
+                writer.WriteLine($"ROT: \n{i.transform.rotation.eulerAngles.x},{i.transform.rotation.eulerAngles.y},{i.transform.rotation.eulerAngles.z}");
+                writer.WriteLine($"SIZ: \n{i.transform.localScale.x},{i.transform.localScale.y},{i.transform.localScale.z}");
+                writer.WriteLine($"COL: \n{i.currentColor.r},{i.currentColor.g},{i.currentColor.b},{i.currentColor.a}");
+                writer.WriteLine($"SKT: \n{i.sketchified.ToString()[0]}");
+                writer.WriteLine($"------------------------------------------");
+            }
+
+            writer.Close();
+
+            Debug.Log("SAVED!!!");
         }
-
-        writer.Close();
-
-        Debug.Log("SAVED!!!");
     }
 }
